poc_backup/requirements.txt
# Core dependencies
llama-index==0.10.12
llama-index-llms-openai==0.1.5
openai==1.12.0
python-dotenv==1.0.0
tavily-python==0.3.0

# Document processing
pypdf==4.0.1
python-docx==1.1.0

# Optional - LlamaParse (comment out if not needed)
llama-parse==0.3.4

# Async support
aiofiles==23.2.1

# Utils
pathlib
shutil